Discounts work properly now.
Needs to be added to the updateBooking method as well in BookingServiceImpl.

Add tests.

+++ src/main/java/com/example/roombooking/dto/BookingLiteDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -18,6 +19,6 @@ public class BookingLiteDTO {
     private int numberOfPeople;
     private LocalDate startDate;
     private LocalDate endDate;
-    private double totalPrice;
+    private BigDecimal totalPrice;
 
 }
